adds support for pull
Adds support for Pull’s
JSON_META
and NEW_LINE output types.
JSON_ARRAY is not yet supported due to a bug.
See the examples in the test folder for usage.
It is recommended that the take() approach demonstrated is used.
Otherwise you must manually perform all status checks, stopping and
unsubscriptions