Ensure we don't stop pulling until subscriptions are finished or failed and emptied
it was previously possible that a push subscription may not be automatically
checked for new data depending on the order in which certain events occurred.
this modification ensures that if we stop checking for data, the subscription's
status is either finished or failed.
if it is failed then the subscription's buffer is checked for data and that data is fetched
until the API indicates that there's nothing more available. At which point the auto
check stops