Prevent race condition between data check and clear
It was possible that when "received" is called it fires doNotify around the same time that `sync` is called.
When this happens, it was possible that between checking if data is null and obtaining the lock to then call `block.clear()`,
that the data object was set and `block.add(new Object())` was called.
If that happened, the `block` queue would get cleared and the code proceeds to block the thread waiting for a response that has
already been received and therefore never unblocks.

This addresses that by making the check for null data and clearing the queue "atomic" so it's not possible for block to have been updated
between checking if data is null and calling `block.clear()`. Whilst, if this happens between the check and call ing `block.take()` the code
immediately proceeds without waiting.